% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar-package.R
\docType{package}
\name{SimilaR-package}
\alias{SimilaR-package}
\alias{SimilaR}
\title{The SimilaR Package}
\description{
See \code{\link{SimilaR_fromDirectory}()} for details.
}
\references{
\emph{\pkg{SimilaR} Package homepage}, \url{http://SimilaR.rexamine.com/}
}
\author{
Maciej Bartoszuk, Marek Gagolewski
}
